requirements.txt
streamlit
sentence-transformers
numpy
pandas
google-generativeai
python-dotenv
scikit-learn
graphviz